requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
pandas
scikit-learn
python-dotenv
google-generativeai
